// client/src/api/model/skillType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Genesys API Documentation
 * Interactive API documentation
 * OpenAPI spec version: 1.0
 */

export type SkillType = typeof SkillType[keyof typeof SkillType];


export const SkillType = {
  General: 'General',
  Magic: 'Magic',
  Combat: 'Combat',
  Social: 'Social',
  Knowledge: 'Knowledge',
} as const;